backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
groq
pydantic
python-multipart
requests
psycopg2-binary

pypdf
python-dotenv
sqlalchemy
aiofiles
bcrypt
PyJWT
